+++ ASI.MGC.FS.Model/rpt_StockReport_Result.cs
@@ -14,7 +14,7 @@ namespace ASI.MGC.FS.Model
     public partial class rpt_StockReport_Result
     {
         public string DESCRIPTION_PM { get; set; }
-        public Nullable<int> CUR_QTY_PM { get; set; }
+        public Nullable<decimal> CUR_QTY_PM { get; set; }
         public Nullable<decimal> RATE_PM { get; set; }
         public string PROD_CODE_PM { get; set; }
         public string STATUS_PM { get; set; }
